Strukturert featuren bedre.
